Create a new regular expression to use in presser routes

Note that presser uses PERL regular expressions.

new_regexp(x)

Arguments

x
String scalar containing a regular expression.

Value

String with class presser_regexp.

Details

As R does not have data type or class for regular expressions, you can use new_regexp() to mark a string as a regular expression, when adding routes.

See also

Examples

new_regexp("^/api/match/(?<pattern>.*)$")

#> <presser_regexp> "^/api/match/(?<pattern>.*)$"