glue based template engine

Use this template engine to create pages with glue templates. See glue::glue() for the syntax.

tmpl_glue(
  sep = "",
  open = "{",
  close = "}",
  na = "NA",
  transformer = NULL,
  trim = TRUE
)

Arguments

sep
Separator used to separate elements.

open
The opening delimiter. Doubling the full delimiter escapes it.

close
The closing delimiter. Doubling the full delimiter escapes it.

na
Value to replace NA values with. If NULL missing values are propagated, that is an NA result will cause NA output. Otherwise the value is replaced by the value of na.

transformer
A function taking three parameters code, envir and data used to transform the output of each block before during or after evaluation.

trim
Whether to trim the input template with glue::trim() or not.

Value

Template function.